apps/polyphemus/pyproject.toml
[project]
name = "polyphemus"
version = "0.2.4"
requires-python = ">=3.10"
dependencies = [
    # FastAPI and server dependencies
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "gunicorn>=21.0.0",

    # Prompt templating and progress bars
    "jinja2>=3.1.6",
    "tqdm>=4.67.1",

    # Data analysis and plotting
    "pandas>=2.2.2",
    "matplotlib>=3.10.0",
    "numpy>=2.2.0",
    "seaborn>=0.13.0",

    # SDK dependency (required for auth and SDK integration)
    "rhesis-sdk[huggingface]",
    # Database dependencies
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    # Backend dependency (brings SDK, Penelope, and core dependencies)
    "rhesis-backend[gpu]",
    # GCS support for model loading from cloud buckets
    "google-cloud-storage>=2.10.0",
]

[project.optional-dependencies]
# Optional: Perspective API for toxicity evaluation
toxicity = [
    "google-api-python-client",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rhesis"]

[tool.uv.sources.rhesis-sdk]
path = "../../sdk"
editable = true

[tool.uv.sources.rhesis-penelope]
path = "../../penelope"
editable = true

[tool.uv.sources.rhesis-backend]
path = "../backend"
editable = true